move Kernel dataclass to rangeify
